Refactor code
Remove the return statement from all getMovesFor... functions
because it is not used.
Remove the stepNumber parameter from all getMovesFor... functions
because it is redundant after making currentColor a global variable.
Rename currentSquareCoords to individualSquareCoords so the naming
of variables is more consistent.
Rename pieceToMove to pieceToMoveName so the naming of variables
is more consistent.